+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
@@ -3,7 +3,7 @@ package nextstep.mvc.controller.tobe;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.Method;
-import nextstep.mvc.exception.tobe.IllegalMethodException;
+import nextstep.mvc.exception.controller.IllegalMethodException;
 import nextstep.mvc.view.ModelAndView;
 
 public class HandlerExecution {
